+++ test/javascript/hologram_test.mjs
@@ -652,8 +652,8 @@ describe("Hologram", () => {
           [
             Type.atom("event"),
             Type.map([
-              [Type.atom("page_x"), Type.integer(1)],
-              [Type.atom("page_y"), Type.integer(2)],
+              [Type.atom("page_x"), Type.float(1)],
+              [Type.atom("page_y"), Type.float(2)],
             ]),
           ],
         ]),
@@ -696,8 +696,8 @@ describe("Hologram", () => {
           [
             Type.atom("event"),
             Type.map([
-              [Type.atom("page_x"), Type.integer(1)],
-              [Type.atom("page_y"), Type.integer(2)],
+              [Type.atom("page_x"), Type.float(1)],
+              [Type.atom("page_y"), Type.float(2)],
             ]),
           ],
         ]),
@@ -735,8 +735,8 @@ describe("Hologram", () => {
           [
             Type.atom("event"),
             Type.map([
-              [Type.atom("page_x"), Type.integer(1)],
-              [Type.atom("page_y"), Type.integer(2)],
+              [Type.atom("page_x"), Type.float(1)],
+              [Type.atom("page_y"), Type.float(2)],
             ]),
           ],
         ]),
